+++ src/main/java/com/demo/server/service/MomentService.java
@@ -1,8 +1,11 @@
 package com.demo.server.service;
 
-import com.demo.server.bean.Moment;
-import com.demo.server.bean.ResultMsg;
+import com.demo.server.bean.*;
+import com.demo.server.dao.AuthDao;
+import com.demo.server.dao.CoupleDao;
 import com.demo.server.dao.MomentDao;
+import com.demo.server.dao.UserDao;
+import com.demo.utils.TokenUtil;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
@@ -17,13 +20,29 @@ import java.util.List;
 public class MomentService {
     @Resource
     private MomentDao momentDao;
+    @Resource
+    private AuthDao authDao;
+    @Resource
+    private UserDao userDao;
+    @Resource
+    private CoupleDao coupleDao;
 
-    public ResultMsg addMoment(Moment moment) {
+    public ResultMsg addMoment(String token, Moment moment) {
         ResultMsg resultMsg = new ResultMsg();
+        if (!checkTokenInvalidation(token, resultMsg)) {
+            return resultMsg;
+        }// token 不合法
+        long uid = TokenUtil.getUidFromToken(token);
+        User user = userDao.selectUserByUid(uid);
+        if (user == null) {
+            resultMsg.setCode(107);
+            return resultMsg;
+        } // 用户不存在
         if (moment == null) {
             resultMsg.setCode(101);
             resultMsg.setMessage("moment is null");
         } else {
+            moment.setUid(uid);
             try {
                 int rowCount = momentDao.insertMoment(moment);
                 System.out.println("rowCount = " + rowCount);
@@ -44,8 +63,17 @@ public class MomentService {
         return resultMsg;
     }
 
-    public ResultMsg updateMoment(Moment moment, long id) {
+    public ResultMsg updateMoment(String token, Moment moment, long id) {
         ResultMsg resultMsg = new ResultMsg();
+        if (!checkTokenInvalidation(token, resultMsg)) {
+            return resultMsg;
+        }// token 不合法
+        long uid = TokenUtil.getUidFromToken(token);
+        User user = userDao.selectUserByUid(uid);
+        if (user == null) {
+            resultMsg.setCode(107);
+            return resultMsg;
+        } // 用户不存在
         if (moment == null) {
             resultMsg.setCode(101);
             resultMsg.setMessage("moment is null");
@@ -69,8 +97,17 @@ public class MomentService {
         return resultMsg;
     }
 
-    public ResultMsg deleteMoment(long id) {
+    public ResultMsg deleteMoment(String token, long id) {
         ResultMsg resultMsg = new ResultMsg();
+        if (!checkTokenInvalidation(token, resultMsg)) {
+            return resultMsg;
+        }// token 不合法
+        long uid = TokenUtil.getUidFromToken(token);
+        User user = userDao.selectUserByUid(uid);
+        if (user == null) {
+            resultMsg.setCode(107);
+            return resultMsg;
+        } // 用户不存在
         try {
             int rowCount = momentDao.deleteMoment(id);
             if (rowCount == 0) {
@@ -85,10 +122,24 @@ public class MomentService {
         return resultMsg;
     }
 
-    public ResultMsg getMoment(int size, long timeCursor) {
+    public ResultMsg getMoment(String token, int size, long timeCursor) {
         ResultMsg resultMsg = new ResultMsg();
+        if (!checkTokenInvalidation(token, resultMsg)) {
+            return resultMsg;
+        }// token 不合法
+        long uid = TokenUtil.getUidFromToken(token);
+        User user = userDao.selectUserByUid(uid);
+        if (user == null) {
+            resultMsg.setCode(107);
+            return resultMsg;
+        } // 用户不存在
         try {
-            List<Moment> momentList = momentDao.selectMoment(size, timeCursor);
+            Couple couple = coupleDao.selectCoupleByLover(uid);
+            long loverId = -1;
+            if (couple != null) {
+                loverId = couple.getLoverAId() == uid ? couple.getLoverBId() : couple.getLoverAId();
+            }
+            List<Moment> momentList = momentDao.selectMoment(size, timeCursor, uid, loverId);
             resultMsg.setCode(100);
             resultMsg.setSize(momentList.size());
             resultMsg.setData(momentList);
@@ -99,10 +150,24 @@ public class MomentService {
         return resultMsg;
     }
 
-    public ResultMsg getAllMoment() {
+    public ResultMsg getAllMoment(String token) {
         ResultMsg resultMsg = new ResultMsg();
+        if (!checkTokenInvalidation(token, resultMsg)) {
+            return resultMsg;
+        }// token 不合法
+        long uid = TokenUtil.getUidFromToken(token);
+        User user = userDao.selectUserByUid(uid);
+        if (user == null) {
+            resultMsg.setCode(107);
+            return resultMsg;
+        } // 用户不存在
         try {
-            List<Moment> momentList = momentDao.selectAllMoment();
+            Couple couple = coupleDao.selectCoupleByLover(uid);
+            long loverId = -1;
+            if (couple != null) {
+                loverId = couple.getLoverAId() == uid ? couple.getLoverBId() : couple.getLoverAId();
+            }
+            List<Moment> momentList = momentDao.selectAllMoment(uid, loverId);
             resultMsg.setCode(100);
             resultMsg.setSize(momentList.size());
             resultMsg.setData(momentList);
@@ -111,5 +176,22 @@ public class MomentService {
             resultMsg.setCode(103);
         }
         return resultMsg;
+    }
+
+    private boolean checkTokenInvalidation(String token, ResultMsg resultMsg) {
+        if (token == null) {
+            resultMsg.setCode(108);
+            return false;
+        }
+        long uid = TokenUtil.getUidFromToken(token);
+        Authorization authorization = authDao.selectAuthByUid(uid);
+        if (authorization == null) {
+            resultMsg.setCode(107);
+            return false;
+        } else if (!token.equals(authorization.getToken())) {
+            resultMsg.setCode(105);
+            return false;
+        }
+        return true;
     }
 }
